+++ packages/video_player/video_player_avfoundation/ios/Classes/messages.g.m
@@ -86,7 +86,7 @@ static id GetNullableObjectAtIndex(NSArray *array, NSInteger key) {
 
 @implementation FLTTextureMessage
 + (instancetype)makeWithTextureId:(NSNumber *)textureId {
-  FLTTextureMessage *pigeonResult = [[FLTTextureMessage alloc] init];
+  FLTTextureMessage* pigeonResult = [[FLTTextureMessage alloc] init];
   pigeonResult.textureId = textureId;
   return pigeonResult;
 }
@@ -107,8 +107,9 @@ static id GetNullableObjectAtIndex(NSArray *array, NSInteger key) {
 @end
 
 @implementation FLTLoopingMessage
-+ (instancetype)makeWithTextureId:(NSNumber *)textureId isLooping:(NSNumber *)isLooping {
-  FLTLoopingMessage *pigeonResult = [[FLTLoopingMessage alloc] init];
++ (instancetype)makeWithTextureId:(NSNumber *)textureId
+    isLooping:(NSNumber *)isLooping {
+  FLTLoopingMessage* pigeonResult = [[FLTLoopingMessage alloc] init];
   pigeonResult.textureId = textureId;
   pigeonResult.isLooping = isLooping;
   return pigeonResult;
@@ -134,7 +135,7 @@ static id GetNullableObjectAtIndex(NSArray *array, NSInteger key) {
 
 @implementation FLTIsSupportedMessage
 + (instancetype)makeWithIsSupported:(NSNumber *)isSupported {
-  FLTIsSupportedMessage *pigeonResult = [[FLTIsSupportedMessage alloc] init];
+  FLTIsSupportedMessage* pigeonResult = [[FLTIsSupportedMessage alloc] init];
   pigeonResult.isSupported = isSupported;
   return pigeonResult;
 }
@@ -156,7 +157,7 @@ static id GetNullableObjectAtIndex(NSArray *array, NSInteger key) {
 
 @implementation FLTIsCacheSupportedMessage
 + (instancetype)makeWithUrl:(NSString *)url {
-  FLTIsCacheSupportedMessage *pigeonResult = [[FLTIsCacheSupportedMessage alloc] init];
+  FLTIsCacheSupportedMessage* pigeonResult = [[FLTIsCacheSupportedMessage alloc] init];
   pigeonResult.url = url;
   return pigeonResult;
 }
@@ -177,8 +178,9 @@ static id GetNullableObjectAtIndex(NSArray *array, NSInteger key) {
 @end
 
 @implementation FLTVolumeMessage
-+ (instancetype)makeWithTextureId:(NSNumber *)textureId volume:(NSNumber *)volume {
-  FLTVolumeMessage *pigeonResult = [[FLTVolumeMessage alloc] init];
++ (instancetype)makeWithTextureId:(NSNumber *)textureId
+    volume:(NSNumber *)volume {
+  FLTVolumeMessage* pigeonResult = [[FLTVolumeMessage alloc] init];
   pigeonResult.textureId = textureId;
   pigeonResult.volume = volume;
   return pigeonResult;
@@ -204,7 +206,7 @@ static id GetNullableObjectAtIndex(NSArray *array, NSInteger key) {
 
 @implementation FLTClearCacheMessage
 + (instancetype)makeWithTextureId:(NSNumber *)textureId {
-  FLTClearCacheMessage *pigeonResult = [[FLTClearCacheMessage alloc] init];
+  FLTClearCacheMessage* pigeonResult = [[FLTClearCacheMessage alloc] init];
   pigeonResult.textureId = textureId;
   return pigeonResult;
 }
@@ -225,8 +227,9 @@ static id GetNullableObjectAtIndex(NSArray *array, NSInteger key) {
 @end
 
 @implementation FLTPlaybackSpeedMessage
-+ (instancetype)makeWithTextureId:(NSNumber *)textureId speed:(NSNumber *)speed {
-  FLTPlaybackSpeedMessage *pigeonResult = [[FLTPlaybackSpeedMessage alloc] init];
++ (instancetype)makeWithTextureId:(NSNumber *)textureId
+    speed:(NSNumber *)speed {
+  FLTPlaybackSpeedMessage* pigeonResult = [[FLTPlaybackSpeedMessage alloc] init];
   pigeonResult.textureId = textureId;
   pigeonResult.speed = speed;
   return pigeonResult;
@@ -251,8 +254,9 @@ static id GetNullableObjectAtIndex(NSArray *array, NSInteger key) {
 @end
 
 @implementation FLTPositionMessage
-+ (instancetype)makeWithTextureId:(NSNumber *)textureId position:(NSNumber *)position {
-  FLTPositionMessage *pigeonResult = [[FLTPositionMessage alloc] init];
++ (instancetype)makeWithTextureId:(NSNumber *)textureId
+    position:(NSNumber *)position {
+  FLTPositionMessage* pigeonResult = [[FLTPositionMessage alloc] init];
   pigeonResult.textureId = textureId;
   pigeonResult.position = position;
   return pigeonResult;
@@ -278,12 +282,12 @@ static id GetNullableObjectAtIndex(NSArray *array, NSInteger key) {
 
 @implementation FLTCreateMessage
 + (instancetype)makeWithAsset:(nullable NSString *)asset
-                          uri:(nullable NSString *)uri
-                  packageName:(nullable NSString *)packageName
-                   formatHint:(nullable NSString *)formatHint
-                  enableCache:(nullable NSNumber *)enableCache
-                  httpHeaders:(NSDictionary<NSString *, NSString *> *)httpHeaders {
-  FLTCreateMessage *pigeonResult = [[FLTCreateMessage alloc] init];
+    uri:(nullable NSString *)uri
+    packageName:(nullable NSString *)packageName
+    formatHint:(nullable NSString *)formatHint
+    enableCache:(nullable NSNumber *)enableCache
+    httpHeaders:(NSDictionary<NSString *, NSString *> *)httpHeaders {
+  FLTCreateMessage* pigeonResult = [[FLTCreateMessage alloc] init];
   pigeonResult.asset = asset;
   pigeonResult.uri = uri;
   pigeonResult.packageName = packageName;
@@ -320,7 +324,7 @@ static id GetNullableObjectAtIndex(NSArray *array, NSInteger key) {
 
 @implementation FLTMixWithOthersMessage
 + (instancetype)makeWithMixWithOthers:(NSNumber *)mixWithOthers {
-  FLTMixWithOthersMessage *pigeonResult = [[FLTMixWithOthersMessage alloc] init];
+  FLTMixWithOthersMessage* pigeonResult = [[FLTMixWithOthersMessage alloc] init];
   pigeonResult.mixWithOthers = mixWithOthers;
   return pigeonResult;
 }
@@ -345,25 +349,25 @@ static id GetNullableObjectAtIndex(NSArray *array, NSInteger key) {
 @implementation FLTAVFoundationVideoPlayerApiCodecReader
 - (nullable id)readValueOfType:(UInt8)type {
   switch (type) {
-    case 128:
+    case 128: 
       return [FLTClearCacheMessage fromList:[self readValue]];
-    case 129:
+    case 129: 
       return [FLTCreateMessage fromList:[self readValue]];
-    case 130:
+    case 130: 
       return [FLTIsCacheSupportedMessage fromList:[self readValue]];
-    case 131:
+    case 131: 
       return [FLTIsSupportedMessage fromList:[self readValue]];
-    case 132:
+    case 132: 
       return [FLTLoopingMessage fromList:[self readValue]];
-    case 133:
+    case 133: 
       return [FLTMixWithOthersMessage fromList:[self readValue]];
-    case 134:
+    case 134: 
       return [FLTPlaybackSpeedMessage fromList:[self readValue]];
-    case 135:
+    case 135: 
       return [FLTPositionMessage fromList:[self readValue]];
-    case 136:
+    case 136: 
       return [FLTTextureMessage fromList:[self readValue]];
-    case 137:
+    case 137: 
       return [FLTVolumeMessage fromList:[self readValue]];
     default:
       return [super readValueOfType:type];
@@ -426,24 +430,21 @@ NSObject<FlutterMessageCodec> *FLTAVFoundationVideoPlayerApiGetCodec(void) {
   static FlutterStandardMessageCodec *sSharedObject = nil;
   static dispatch_once_t sPred = 0;
   dispatch_once(&sPred, ^{
-    FLTAVFoundationVideoPlayerApiCodecReaderWriter *readerWriter =
-        [[FLTAVFoundationVideoPlayerApiCodecReaderWriter alloc] init];
+    FLTAVFoundationVideoPlayerApiCodecReaderWriter *readerWriter = [[FLTAVFoundationVideoPlayerApiCodecReaderWriter alloc] init];
     sSharedObject = [FlutterStandardMessageCodec codecWithReaderWriter:readerWriter];
   });
   return sSharedObject;
 }
 
-void FLTAVFoundationVideoPlayerApiSetup(id<FlutterBinaryMessenger> binaryMessenger,
-                                        NSObject<FLTAVFoundationVideoPlayerApi> *api) {
+void FLTAVFoundationVideoPlayerApiSetup(id<FlutterBinaryMessenger> binaryMessenger, NSObject<FLTAVFoundationVideoPlayerApi> *api) {
   {
-    FlutterBasicMessageChannel *channel = [[FlutterBasicMessageChannel alloc]
-           initWithName:@"dev.flutter.pigeon.AVFoundationVideoPlayerApi.initialize"
+    FlutterBasicMessageChannel *channel =
+      [[FlutterBasicMessageChannel alloc]
+        initWithName:@"dev.flutter.pigeon.AVFoundationVideoPlayerApi.initialize"
         binaryMessenger:binaryMessenger
-                  codec:FLTAVFoundationVideoPlayerApiGetCodec()];
+        codec:FLTAVFoundationVideoPlayerApiGetCodec()];
     if (api) {
-      NSCAssert([api respondsToSelector:@selector(initialize:)],
-                @"FLTAVFoundationVideoPlayerApi api (%@) doesn't respond to @selector(initialize:)",
-                api);
+      NSCAssert([api respondsToSelector:@selector(initialize:)], @"FLTAVFoundationVideoPlayerApi api (%@) doesn't respond to @selector(initialize:)", api);
       [channel setMessageHandler:^(id _Nullable message, FlutterReply callback) {
         FlutterError *error;
         [api initialize:&error];
@@ -454,15 +455,13 @@ void FLTAVFoundationVideoPlayerApiSetup(id<FlutterBinaryMessenger> binaryMesseng
     }
   }
   {
-    FlutterBasicMessageChannel *channel = [[FlutterBasicMessageChannel alloc]
-           initWithName:@"dev.flutter.pigeon.AVFoundationVideoPlayerApi.create"
+    FlutterBasicMessageChannel *channel =
+      [[FlutterBasicMessageChannel alloc]
+        initWithName:@"dev.flutter.pigeon.AVFoundationVideoPlayerApi.create"
         binaryMessenger:binaryMessenger
-                  codec:FLTAVFoundationVideoPlayerApiGetCodec()];
+        codec:FLTAVFoundationVideoPlayerApiGetCodec()];
     if (api) {
-      NSCAssert(
-          [api respondsToSelector:@selector(create:error:)],
-          @"FLTAVFoundationVideoPlayerApi api (%@) doesn't respond to @selector(create:error:)",
-          api);
+      NSCAssert([api respondsToSelector:@selector(create:error:)], @"FLTAVFoundationVideoPlayerApi api (%@) doesn't respond to @selector(create:error:)", api);
       [channel setMessageHandler:^(id _Nullable message, FlutterReply callback) {
         NSArray *args = message;
         FLTCreateMessage *arg_msg = GetNullableObjectAtIndex(args, 0);
@@ -475,15 +474,13 @@ void FLTAVFoundationVideoPlayerApiSetup(id<FlutterBinaryMessenger> binaryMesseng
     }
   }
   {
-    FlutterBasicMessageChannel *channel = [[FlutterBasicMessageChannel alloc]
-           initWithName:@"dev.flutter.pigeon.AVFoundationVideoPlayerApi.dispose"
+    FlutterBasicMessageChannel *channel =
+      [[FlutterBasicMessageChannel alloc]
+        initWithName:@"dev.flutter.pigeon.AVFoundationVideoPlayerApi.dispose"
         binaryMessenger:binaryMessenger
-                  codec:FLTAVFoundationVideoPlayerApiGetCodec()];
+        codec:FLTAVFoundationVideoPlayerApiGetCodec()];
     if (api) {
-      NSCAssert(
-          [api respondsToSelector:@selector(dispose:error:)],
-          @"FLTAVFoundationVideoPlayerApi api (%@) doesn't respond to @selector(dispose:error:)",
-          api);
+      NSCAssert([api respondsToSelector:@selector(dispose:error:)], @"FLTAVFoundationVideoPlayerApi api (%@) doesn't respond to @selector(dispose:error:)", api);
       [channel setMessageHandler:^(id _Nullable message, FlutterReply callback) {
         NSArray *args = message;
         FLTTextureMessage *arg_msg = GetNullableObjectAtIndex(args, 0);
@@ -496,15 +493,13 @@ void FLTAVFoundationVideoPlayerApiSetup(id<FlutterBinaryMessenger> binaryMesseng
     }
   }
   {
-    FlutterBasicMessageChannel *channel = [[FlutterBasicMessageChannel alloc]
-           initWithName:@"dev.flutter.pigeon.AVFoundationVideoPlayerApi.setLooping"
+    FlutterBasicMessageChannel *channel =
+      [[FlutterBasicMessageChannel alloc]
+        initWithName:@"dev.flutter.pigeon.AVFoundationVideoPlayerApi.setLooping"
         binaryMessenger:binaryMessenger
-                  codec:FLTAVFoundationVideoPlayerApiGetCodec()];
+        codec:FLTAVFoundationVideoPlayerApiGetCodec()];
     if (api) {
-      NSCAssert(
-          [api respondsToSelector:@selector(setLooping:error:)],
-          @"FLTAVFoundationVideoPlayerApi api (%@) doesn't respond to @selector(setLooping:error:)",
-          api);
+      NSCAssert([api respondsToSelector:@selector(setLooping:error:)], @"FLTAVFoundationVideoPlayerApi api (%@) doesn't respond to @selector(setLooping:error:)", api);
       [channel setMessageHandler:^(id _Nullable message, FlutterReply callback) {
         NSArray *args = message;
         FLTLoopingMessage *arg_msg = GetNullableObjectAtIndex(args, 0);
@@ -517,15 +512,13 @@ void FLTAVFoundationVideoPlayerApiSetup(id<FlutterBinaryMessenger> binaryMesseng
     }
   }
   {
-    FlutterBasicMessageChannel *channel = [[FlutterBasicMessageChannel alloc]
-           initWithName:@"dev.flutter.pigeon.AVFoundationVideoPlayerApi.clearCache"
+    FlutterBasicMessageChannel *channel =
+      [[FlutterBasicMessageChannel alloc]
+        initWithName:@"dev.flutter.pigeon.AVFoundationVideoPlayerApi.clearCache"
         binaryMessenger:binaryMessenger
-                  codec:FLTAVFoundationVideoPlayerApiGetCodec()];
+        codec:FLTAVFoundationVideoPlayerApiGetCodec()];
     if (api) {
-      NSCAssert(
-          [api respondsToSelector:@selector(clearCache:error:)],
-          @"FLTAVFoundationVideoPlayerApi api (%@) doesn't respond to @selector(clearCache:error:)",
-          api);
+      NSCAssert([api respondsToSelector:@selector(clearCache:error:)], @"FLTAVFoundationVideoPlayerApi api (%@) doesn't respond to @selector(clearCache:error:)", api);
       [channel setMessageHandler:^(id _Nullable message, FlutterReply callback) {
         NSArray *args = message;
         FLTClearCacheMessage *arg_msg = GetNullableObjectAtIndex(args, 0);
@@ -538,15 +531,13 @@ void FLTAVFoundationVideoPlayerApiSetup(id<FlutterBinaryMessenger> binaryMesseng
     }
   }
   {
-    FlutterBasicMessageChannel *channel = [[FlutterBasicMessageChannel alloc]
-           initWithName:@"dev.flutter.pigeon.AVFoundationVideoPlayerApi.setVolume"
+    FlutterBasicMessageChannel *channel =
+      [[FlutterBasicMessageChannel alloc]
+        initWithName:@"dev.flutter.pigeon.AVFoundationVideoPlayerApi.setVolume"
         binaryMessenger:binaryMessenger
-                  codec:FLTAVFoundationVideoPlayerApiGetCodec()];
+        codec:FLTAVFoundationVideoPlayerApiGetCodec()];
     if (api) {
-      NSCAssert(
-          [api respondsToSelector:@selector(setVolume:error:)],
-          @"FLTAVFoundationVideoPlayerApi api (%@) doesn't respond to @selector(setVolume:error:)",
-          api);
+      NSCAssert([api respondsToSelector:@selector(setVolume:error:)], @"FLTAVFoundationVideoPlayerApi api (%@) doesn't respond to @selector(setVolume:error:)", api);
       [channel setMessageHandler:^(id _Nullable message, FlutterReply callback) {
         NSArray *args = message;
         FLTVolumeMessage *arg_msg = GetNullableObjectAtIndex(args, 0);
@@ -559,16 +550,13 @@ void FLTAVFoundationVideoPlayerApiSetup(id<FlutterBinaryMessenger> binaryMesseng
     }
   }
   {
-    FlutterBasicMessageChannel *channel = [[FlutterBasicMessageChannel alloc]
-           initWithName:
-               @"dev.flutter.pigeon.AVFoundationVideoPlayerApi.isCacheSupportedForNetworkMedia"
+    FlutterBasicMessageChannel *channel =
+      [[FlutterBasicMessageChannel alloc]
+        initWithName:@"dev.flutter.pigeon.AVFoundationVideoPlayerApi.isCacheSupportedForNetworkMedia"
         binaryMessenger:binaryMessenger
-                  codec:FLTAVFoundationVideoPlayerApiGetCodec()];
+        codec:FLTAVFoundationVideoPlayerApiGetCodec()];
     if (api) {
-      NSCAssert([api respondsToSelector:@selector(isCacheSupportedForNetworkMedia:error:)],
-                @"FLTAVFoundationVideoPlayerApi api (%@) doesn't respond to "
-                @"@selector(isCacheSupportedForNetworkMedia:error:)",
-                api);
+      NSCAssert([api respondsToSelector:@selector(isCacheSupportedForNetworkMedia:error:)], @"FLTAVFoundationVideoPlayerApi api (%@) doesn't respond to @selector(isCacheSupportedForNetworkMedia:error:)", api);
       [channel setMessageHandler:^(id _Nullable message, FlutterReply callback) {
         NSArray *args = message;
         FLTIsCacheSupportedMessage *arg_msg = GetNullableObjectAtIndex(args, 0);
@@ -581,15 +569,13 @@ void FLTAVFoundationVideoPlayerApiSetup(id<FlutterBinaryMessenger> binaryMesseng
     }
   }
   {
-    FlutterBasicMessageChannel *channel = [[FlutterBasicMessageChannel alloc]
-           initWithName:@"dev.flutter.pigeon.AVFoundationVideoPlayerApi.setPlaybackSpeed"
+    FlutterBasicMessageChannel *channel =
+      [[FlutterBasicMessageChannel alloc]
+        initWithName:@"dev.flutter.pigeon.AVFoundationVideoPlayerApi.setPlaybackSpeed"
         binaryMessenger:binaryMessenger
-                  codec:FLTAVFoundationVideoPlayerApiGetCodec()];
+        codec:FLTAVFoundationVideoPlayerApiGetCodec()];
     if (api) {
-      NSCAssert([api respondsToSelector:@selector(setPlaybackSpeed:error:)],
-                @"FLTAVFoundationVideoPlayerApi api (%@) doesn't respond to "
-                @"@selector(setPlaybackSpeed:error:)",
-                api);
+      NSCAssert([api respondsToSelector:@selector(setPlaybackSpeed:error:)], @"FLTAVFoundationVideoPlayerApi api (%@) doesn't respond to @selector(setPlaybackSpeed:error:)", api);
       [channel setMessageHandler:^(id _Nullable message, FlutterReply callback) {
         NSArray *args = message;
         FLTPlaybackSpeedMessage *arg_msg = GetNullableObjectAtIndex(args, 0);
@@ -602,14 +588,13 @@ void FLTAVFoundationVideoPlayerApiSetup(id<FlutterBinaryMessenger> binaryMesseng
     }
   }
   {
-    FlutterBasicMessageChannel *channel = [[FlutterBasicMessageChannel alloc]
-           initWithName:@"dev.flutter.pigeon.AVFoundationVideoPlayerApi.play"
+    FlutterBasicMessageChannel *channel =
+      [[FlutterBasicMessageChannel alloc]
+        initWithName:@"dev.flutter.pigeon.AVFoundationVideoPlayerApi.play"
         binaryMessenger:binaryMessenger
-                  codec:FLTAVFoundationVideoPlayerApiGetCodec()];
+        codec:FLTAVFoundationVideoPlayerApiGetCodec()];
     if (api) {
-      NSCAssert([api respondsToSelector:@selector(play:error:)],
-                @"FLTAVFoundationVideoPlayerApi api (%@) doesn't respond to @selector(play:error:)",
-                api);
+      NSCAssert([api respondsToSelector:@selector(play:error:)], @"FLTAVFoundationVideoPlayerApi api (%@) doesn't respond to @selector(play:error:)", api);
       [channel setMessageHandler:^(id _Nullable message, FlutterReply callback) {
         NSArray *args = message;
         FLTTextureMessage *arg_msg = GetNullableObjectAtIndex(args, 0);
@@ -622,15 +607,13 @@ void FLTAVFoundationVideoPlayerApiSetup(id<FlutterBinaryMessenger> binaryMesseng
     }
   }
   {
-    FlutterBasicMessageChannel *channel = [[FlutterBasicMessageChannel alloc]
-           initWithName:@"dev.flutter.pigeon.AVFoundationVideoPlayerApi.position"
+    FlutterBasicMessageChannel *channel =
+      [[FlutterBasicMessageChannel alloc]
+        initWithName:@"dev.flutter.pigeon.AVFoundationVideoPlayerApi.position"
         binaryMessenger:binaryMessenger
-                  codec:FLTAVFoundationVideoPlayerApiGetCodec()];
+        codec:FLTAVFoundationVideoPlayerApiGetCodec()];
     if (api) {
-      NSCAssert(
-          [api respondsToSelector:@selector(position:error:)],
-          @"FLTAVFoundationVideoPlayerApi api (%@) doesn't respond to @selector(position:error:)",
-          api);
+      NSCAssert([api respondsToSelector:@selector(position:error:)], @"FLTAVFoundationVideoPlayerApi api (%@) doesn't respond to @selector(position:error:)", api);
       [channel setMessageHandler:^(id _Nullable message, FlutterReply callback) {
         NSArray *args = message;
         FLTTextureMessage *arg_msg = GetNullableObjectAtIndex(args, 0);
@@ -643,37 +626,32 @@ void FLTAVFoundationVideoPlayerApiSetup(id<FlutterBinaryMessenger> binaryMesseng
     }
   }
   {
-    FlutterBasicMessageChannel *channel = [[FlutterBasicMessageChannel alloc]
-           initWithName:@"dev.flutter.pigeon.AVFoundationVideoPlayerApi.seekTo"
+    FlutterBasicMessageChannel *channel =
+      [[FlutterBasicMessageChannel alloc]
+        initWithName:@"dev.flutter.pigeon.AVFoundationVideoPlayerApi.seekTo"
         binaryMessenger:binaryMessenger
-                  codec:FLTAVFoundationVideoPlayerApiGetCodec()];
+        codec:FLTAVFoundationVideoPlayerApiGetCodec()];
     if (api) {
-      NSCAssert([api respondsToSelector:@selector(seekTo:completion:)],
-                @"FLTAVFoundationVideoPlayerApi api (%@) doesn't respond to "
-                @"@selector(seekTo:completion:)",
-                api);
+      NSCAssert([api respondsToSelector:@selector(seekTo:completion:)], @"FLTAVFoundationVideoPlayerApi api (%@) doesn't respond to @selector(seekTo:completion:)", api);
       [channel setMessageHandler:^(id _Nullable message, FlutterReply callback) {
         NSArray *args = message;
         FLTPositionMessage *arg_msg = GetNullableObjectAtIndex(args, 0);
-        [api seekTo:arg_msg
-            completion:^(FlutterError *_Nullable error) {
-              callback(wrapResult(nil, error));
-            }];
+        [api seekTo:arg_msg completion:^(FlutterError *_Nullable error) {
+          callback(wrapResult(nil, error));
+        }];
       }];
     } else {
       [channel setMessageHandler:nil];
     }
   }
   {
-    FlutterBasicMessageChannel *channel = [[FlutterBasicMessageChannel alloc]
-           initWithName:@"dev.flutter.pigeon.AVFoundationVideoPlayerApi.pause"
+    FlutterBasicMessageChannel *channel =
+      [[FlutterBasicMessageChannel alloc]
+        initWithName:@"dev.flutter.pigeon.AVFoundationVideoPlayerApi.pause"
         binaryMessenger:binaryMessenger
-                  codec:FLTAVFoundationVideoPlayerApiGetCodec()];
+        codec:FLTAVFoundationVideoPlayerApiGetCodec()];
     if (api) {
-      NSCAssert(
-          [api respondsToSelector:@selector(pause:error:)],
-          @"FLTAVFoundationVideoPlayerApi api (%@) doesn't respond to @selector(pause:error:)",
-          api);
+      NSCAssert([api respondsToSelector:@selector(pause:error:)], @"FLTAVFoundationVideoPlayerApi api (%@) doesn't respond to @selector(pause:error:)", api);
       [channel setMessageHandler:^(id _Nullable message, FlutterReply callback) {
         NSArray *args = message;
         FLTTextureMessage *arg_msg = GetNullableObjectAtIndex(args, 0);
@@ -686,15 +664,13 @@ void FLTAVFoundationVideoPlayerApiSetup(id<FlutterBinaryMessenger> binaryMesseng
     }
   }
   {
-    FlutterBasicMessageChannel *channel = [[FlutterBasicMessageChannel alloc]
-           initWithName:@"dev.flutter.pigeon.AVFoundationVideoPlayerApi.setMixWithOthers"
+    FlutterBasicMessageChannel *channel =
+      [[FlutterBasicMessageChannel alloc]
+        initWithName:@"dev.flutter.pigeon.AVFoundationVideoPlayerApi.setMixWithOthers"
         binaryMessenger:binaryMessenger
-                  codec:FLTAVFoundationVideoPlayerApiGetCodec()];
+        codec:FLTAVFoundationVideoPlayerApiGetCodec()];
     if (api) {
-      NSCAssert([api respondsToSelector:@selector(setMixWithOthers:error:)],
-                @"FLTAVFoundationVideoPlayerApi api (%@) doesn't respond to "
-                @"@selector(setMixWithOthers:error:)",
-                api);
+      NSCAssert([api respondsToSelector:@selector(setMixWithOthers:error:)], @"FLTAVFoundationVideoPlayerApi api (%@) doesn't respond to @selector(setMixWithOthers:error:)", api);
       [channel setMessageHandler:^(id _Nullable message, FlutterReply callback) {
         NSArray *args = message;
         FLTMixWithOthersMessage *arg_msg = GetNullableObjectAtIndex(args, 0);