+++ packages/video_player/video_player_platform_interface/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Adds caching functionality for network video.
 * Updates minimum supported SDK version to Flutter 3.3/Dart 2.18.
+* Adds caching functionality for network video.
 
 ## 6.1.0